Added new test type
There is currently an issue that prevents Etheno from identifying newly deployed transactions that happen through function calls. This is because there is no way for Etheno (or even Ganache) to discern when a transaction ends up creating a contract during its execution or not. This test case will be used for debugging and building this feature.